Fix compare line_number to eline which is not set
If a scope was not yet closed eline is default to None.
For checking if a certain line is within a given scope
we first have to check if the scope is already closed
and only than compare the given line number against eline.
If a scope is not closed the line is always within the
scope as long as the line comes after the start line of
the scope.