README: add fix for possible installation error
Since commit 2e6f2f4 (Do not install future and argparse unless
necessary, 2019-09-18), the syntax used in the 'install_requires' field
of setup.py is not compatible with very old versions of setuptools.

Add a note to the README to direct users to upgrade setuptools if they
get this error when installing fortran-language-server.